Added basic database functions.
