+++ run
@@ -3,4 +3,12 @@ export FLASK_APP=server.py
 # Change this for demos, otherwise there will be debugger stack traces written to the web page
 export FLASK_ENV=development
 
-cd src && flask run && cd ..
+cd src
+
+if [ -z $1 ]; then
+	flask run
+elif [ $1 = "debug" ]; then
+	flask shell
+fi
+
+cd ..
