+++ src/schema.sql
@@ -1,14 +1,13 @@
 -- uncomment this line if you want to refresh your database on next run
 -- DROP TABLE IF EXISTS paper;
 CREATE TABLE IF NOT EXISTS paper (
-  `sequence_id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-  `paper_id` VARCHAR(255) NOT NULL,
+  `paper_id` VARCHAR(255) PRIMARY KEY NOT NULL,
   `has_bib` BOOLEAN NOT NULL DEFAULT 0,
   `has_tex` BOOLEAN NOT NULL DEFAULT 0,
   `title_bib` VARCHAR(255),
   `title_tex` VARCHAR(255),
   `author_bib` VARCHAR(255),
-  `author_tex` VARCHAR(255),
+  `author_tex` VARCHAR(512),
   `year` INTEGER,
   `month` VARCHAR(255),
   `volume` VARCHAR(255),
