+++ src/server.py
@@ -1,6 +1,7 @@
-import os, webbrowser, sqlite3
+import os, webbrowser
 from flask import *
 from blueprints import ref_info, lookup_paper
+from db import *
 
 # generate the doxygen configuration file
 # os.system("doxygen -g ../data/out/doxygen/config/Doxyfile")
@@ -10,15 +11,6 @@ from blueprints import ref_info, lookup_paper
 
 # open HTML in browser
 # webbrowser.open("../data/out/doxygen/html/index.html")
-
-DATABASE = 'citation_graph.db'
-QUERY_GET_PAPER_BY_ID = 'SELECT * FROM paper WHERE paper_id = ?' # working
-QUERY_INSERT_PAPER = 'INSERT INTO paper (paper_id, has_bib, has_tex) VALUES (?, ?, ?)' # working
-QUERY_UPDATE_ATTRIBUTE_ABSTRACT = 'UPDATE paper SET abstract = ? WHERE paper_id = ?' # working
-QUERY_GET_PAPERS_BY_ATTRIBUTE = 'SELECT * FROM paper WHERE ? = ?' # untested
-QUERY_GET_ATTRIBUTE_BY_ID = 'SELECT ? FROM paper WHERE paper_id = ?' # untested
-QUERY_GET_CITATIONS_BY_ID = 'SELECT reference_paper_id FROM citation WHERE paper_id = ?' # untested
-QUERY_INSERT_CITATION = 'INSERT INTO citation (paper_id, reference_paper_id) VALUES (?, ?)' # untested
 
 def create_app(test_config=None):
 	if os.path.abspath(os.curdir).find(" ") != -1:
@@ -31,6 +23,7 @@ def create_app(test_config=None):
 
 	app.config.from_object("config.Config")
 	app.config['USER_PAPERS'] = os.path.join(app.instance_path,'user_files', 'papers')
+	app.config['DATABASE'] = 'citation_graph.db'
 
 	if test_config is None:
 		app.config.from_pyfile('config.py', silent=True)
@@ -54,41 +47,7 @@ def create_app(test_config=None):
 	app.register_blueprint(ref_info.bp)
 	app.register_blueprint(lookup_paper.bp)
 
-	def get_db():
-		db = getattr(g, '_database', None)
-		if db is None:
-			db = g._database = sqlite3.connect(DATABASE)
-		return db
-
-	def init_db(app):
-		with app.app_context():
-			db = get_db()
-			with app.open_resource('schema.sql', mode='r') as f:
-				db.cursor().executescript(f.read())
-			db.commit()
-
 	init_db(app)
-
-	# for getting results from database
-	# param one True if you only want one result
-	def query_db(query, args=(), one=False):
-		cur = get_db().execute(query, args)
-		rv = cur.fetchall()
-		cur.close()
-		return (rv[0] if rv else None) if one else rv
-	
-	# for executing a query on the database
-	def execute_db(query, args=()):
-		try:
-			with sqlite3.connect(DATABASE) as con:
-				cur = con.cursor()
-				cur.execute(query, args)
-				con.commit()
-		except:
-			print('Failed to execute query.')
-			con.rollback()
-		finally:
-			con.close()
 
 	# the commented out lines allow the data to come from the front-end
 	# @app.route('/exampleAddPaper', methods = ['POST', 'GET'])
@@ -120,7 +79,7 @@ def create_app(test_config=None):
 		else:
 			print('Retrieved whole paper, can access attributes like this:' + paper[1])
 		return render_template('index.html') # render whatever page you want, index is just a placeholder
-	
+
 	# the commented out lines allow the data to come from the front-end
 	# @app.route('/exampleUpdateAttribute', methods = ['POST', 'GET'])
 	@app.route('/exampleUpdateAttribute')
@@ -132,4 +91,3 @@ def create_app(test_config=None):
 		return render_template('index.html') # render whatever page you want, index is just a placeholder
 
 	return app
-
